fix: overhaul golangci-lint config and fix all lint errors
Switch .golangci.yaml from `default: all` (111 linters) to `default: none`
with 62 explicitly enabled linters. Fix all reported issues across ~550 files:
- gocritic: if-else chains → switch, unlambda, assignOp, elseif
- unused: remove dead code, nolint for build-tagged code
- unconvert: remove unnecessary type conversions
- wastedassign: fix wasted initial assignments
- staticcheck: apply De Morgan's law simplifications
- perfsprint: use errors.New where no formatting needed
- gci: fix import grouping
- errchkjson: add error checks for json.Marshal
- forcetypeassert: add checked type assertions
- gocheckcompilerdirectives: remove invalid //go:fix directive
- nlreturn: add blank lines before returns
- Code formatting via gci/gofmt/goimports

+++ internal/storage/readstore/metrics.go
@@ -55,9 +55,10 @@ func (s *Store) RegisterMetrics(m metric.Meter) (metric.Registration, error) {
 		o.ObserveInt64(freelistPages, int64(stats.FreePageN), metric.WithAttributes(attribute.String("type", "free")))
 		o.ObserveInt64(freelistPages, int64(stats.PendingPageN), metric.WithAttributes(attribute.String("type", "pending")))
 		o.ObserveInt64(freelistBytes, int64(stats.FreeAlloc))
-		o.ObserveInt64(txPageReads, int64(stats.TxStats.PageCount))
+		o.ObserveInt64(txPageReads, stats.TxStats.PageCount)
+
 		if pageSize > 0 {
-			o.ObserveInt64(txPageWrites, int64(stats.TxStats.PageAlloc)/pageSize)
+			o.ObserveInt64(txPageWrites, stats.TxStats.PageAlloc/pageSize)
 		}
 
 		return nil
